Extract StorageConfig from store.rs into separate module.
Continuing to decompose the structure of store.rs.
