+++ src/store/store.rs
@@ -35,9 +35,9 @@ use super::enqueue_batcher::EnqueueBatcher;
 use super::group_committer::GroupCommitter;
 use super::ready_index::ReadyIndex;
 use super::scheduled_index::ScheduledIndex;
+use super::storage_config::StorageConfig;
 use super::types::{
-    BackoffConfig, CommitMode, EnvConfigError, Job, JobStatus, StoreError, UniqueConstraint,
-    UniqueWhile,
+    BackoffConfig, CommitMode, Job, JobStatus, StoreError, UniqueConstraint, UniqueWhile,
 };
 
 /// Events broadcast by the store when job state changes.
@@ -280,242 +280,6 @@ impl Keyspaces {
             )))),
         }
     }
-}
-
-/// Tuning parameters for the underlying LSM storage engine.
-///
-/// Job queue workloads are high-churn: jobs are enqueued, processed,
-/// and deleted rapidly, generating many tombstones across all keyspaces.
-/// The fjall defaults (64 MiB tables, 512 MiB journal) are tuned for
-/// larger, read-heavy workloads. These defaults are sized down for
-/// faster tombstone reclamation and tighter disk usage.
-#[derive(Debug, Clone)]
-pub struct StorageConfig {
-    /// Block cache capacity in bytes. Shared across all keyspaces.
-    /// Increase when the in-flight set of filter + index + hot data blocks
-    /// exceeds the default. Recommended: 20–25 % of available memory.
-    pub cache_size: u64,
-
-    /// Target size for SST files in the `data` keyspace (jobs, payloads, errors).
-    /// The memtable is sized to match so flushes produce appropriately-sized
-    /// L0 tables.
-    pub data_table_size: u64,
-
-    /// Target size for SST files in the `index` keyspace (status, queue, type,
-    /// purge-at indexes). These store tiny entries so smaller tables are
-    /// appropriate. The memtable is sized to match.
-    pub index_table_size: u64,
-
-    /// Maximum total size of the WAL (journal) on disk. The journal is
-    /// shared across all keyspaces and can only be reclaimed once every
-    /// keyspace referenced in it has flushed. The minimum is 64 MiB.
-    pub journal_size: u64,
-
-    /// Number of L0 files that triggers compaction into L1. Lower values
-    /// mean more frequent compaction (less L0 buildup, faster reads after
-    /// recovery) but more write amplification during steady state.
-    pub l0_threshold: u8,
-
-    /// Default retention period for completed jobs (milliseconds).
-    /// 0 means completed jobs are logically invisible immediately.
-    pub default_completed_retention_ms: u64,
-
-    /// Default retention period for dead jobs (milliseconds).
-    pub default_dead_retention_ms: u64,
-
-    /// Default maximum retries before a failed job is killed.
-    pub default_retry_limit: u32,
-
-    /// Default backoff config applied to jobs that don't specify one.
-    pub default_backoff: BackoffConfig,
-
-    /// Default commit mode for most operations.
-    pub default_commit_mode: CommitMode,
-
-    /// Optional per-operation commit mode for enqueue. When `None`,
-    /// enqueue inherits `default_commit_mode`.
-    pub enqueue_commit_mode: Option<CommitMode>,
-
-    /// Maximum number of single-job enqueues coalesced into one write
-    /// transaction by the server-side auto-batcher. Also the bounded
-    /// channel capacity for in-flight enqueue requests, so it caps both
-    /// per-batch size and total in-flight load. Does NOT apply to
-    /// explicit `enqueue_bulk` calls, which run their own tx.
-    pub enqueue_batch_size: usize,
-
-    /// Maximum number of completion (ack) requests coalesced into one
-    /// write transaction. Op-count bounded — a bulk-ack of N jobs
-    /// still counts as one op. Same shape as `enqueue_batch_size`.
-    pub complete_batch_size: usize,
-
-    /// After a bulk mutation affects at least this many records, force a
-    /// full LSM compaction to reclaim tombstones. Without this, leveled
-    /// compaction can leave large pockets of garbage in the upper levels
-    /// for a long time on quiet databases. Set to 0 to disable.
-    pub auto_compact_threshold: u64,
-}
-
-/// Default block cache capacity (256 MiB).
-pub const DEFAULT_CACHE_SIZE: u64 = 256 * 1024 * 1024;
-
-/// Default target size for data SST files (64 MiB).
-pub const DEFAULT_DATA_TABLE_SIZE: u64 = 64 * 1024 * 1024;
-
-/// Default target size for index SST files (8 MiB).
-pub const DEFAULT_INDEX_TABLE_SIZE: u64 = 8 * 1024 * 1024;
-
-/// Default maximum journal (WAL) size on disk (64 MiB minimum).
-pub const DEFAULT_JOURNAL_SIZE: u64 = 64 * 1024 * 1024;
-
-/// Default L0 compaction threshold.
-pub const DEFAULT_L0_THRESHOLD: u8 = 4;
-
-/// Default maximum retries before a failed job is killed.
-pub const DEFAULT_RETRY_LIMIT: u32 = 25;
-
-/// Default backoff exponent (power curve steepness).
-pub const DEFAULT_BACKOFF_EXPONENT: f32 = 4.0;
-
-/// Default backoff base delay in milliseconds.
-pub const DEFAULT_BACKOFF_BASE_MS: u32 = 15_000;
-
-/// Default backoff jitter in milliseconds (max random ms per attempt multiplier).
-pub const DEFAULT_BACKOFF_JITTER_MS: u32 = 30_000;
-
-/// Default retention period for completed jobs (milliseconds).
-/// 0 means completed jobs are logically invisible immediately.
-pub const DEFAULT_COMPLETED_RETENTION_MS: u64 = 0;
-
-/// Default retention period for dead jobs (milliseconds). 7 days.
-pub const DEFAULT_DEAD_RETENTION_MS: u64 = 604_800_000;
-
-/// Default threshold for auto-compacting after a bulk mutation. Bulk
-/// operations that mutate at least this many records trigger a full
-/// compaction once they commit.
-pub const DEFAULT_AUTO_COMPACT_THRESHOLD: u64 = 10_000;
-
-/// Default maximum number of concurrent single-job enqueues coalesced
-/// into one auto-batched commit. Also the bounded channel capacity.
-pub const DEFAULT_ENQUEUE_BATCH_SIZE: usize = 1000;
-
-/// Default maximum number of concurrent completion (ack) requests
-/// coalesced into one auto-batched commit. Also the bounded channel
-/// capacity.
-pub const DEFAULT_COMPLETE_BATCH_SIZE: usize = 1000;
-
-impl Default for StorageConfig {
-    fn default() -> Self {
-        Self {
-            cache_size: DEFAULT_CACHE_SIZE,
-            data_table_size: DEFAULT_DATA_TABLE_SIZE,
-            index_table_size: DEFAULT_INDEX_TABLE_SIZE,
-            journal_size: DEFAULT_JOURNAL_SIZE,
-            l0_threshold: DEFAULT_L0_THRESHOLD,
-            default_completed_retention_ms: DEFAULT_COMPLETED_RETENTION_MS,
-            default_dead_retention_ms: DEFAULT_DEAD_RETENTION_MS,
-            default_retry_limit: DEFAULT_RETRY_LIMIT,
-            default_backoff: BackoffConfig {
-                exponent: DEFAULT_BACKOFF_EXPONENT,
-                base_ms: DEFAULT_BACKOFF_BASE_MS,
-                jitter_ms: DEFAULT_BACKOFF_JITTER_MS,
-            },
-            default_commit_mode: CommitMode::default(),
-            enqueue_commit_mode: None,
-            enqueue_batch_size: DEFAULT_ENQUEUE_BATCH_SIZE,
-            complete_batch_size: DEFAULT_COMPLETE_BATCH_SIZE,
-            auto_compact_threshold: DEFAULT_AUTO_COMPACT_THRESHOLD,
-        }
-    }
-}
-
-impl StorageConfig {
-    /// Build a `StorageConfig` from environment variables, falling back to
-    /// defaults for any that are unset. Returns an error if a variable is
-    /// set but cannot be parsed.
-    ///
-    /// | Variable                    | Field              |
-    /// |-----------------------------|--------------------|
-    /// | `ZIZQ_CACHE_SIZE`         | `cache_size`       |
-    /// | `ZIZQ_DATA_TABLE_SIZE`    | `data_table_size`  |
-    /// | `ZIZQ_INDEX_TABLE_SIZE`   | `index_table_size` |
-    /// | `ZIZQ_JOURNAL_SIZE`       | `journal_size`     |
-    /// | `ZIZQ_L0_THRESHOLD`       | `l0_threshold`     |
-    /// | `ZIZQ_DEFAULT_COMMIT_MODE` | `default_commit_mode` |
-    /// | `ZIZQ_ENQUEUE_COMMIT_MODE` | `enqueue_commit_mode` |
-    /// | `ZIZQ_ENQUEUE_BATCH_SIZE` | `enqueue_batch_size` |
-    /// | `ZIZQ_COMPLETE_BATCH_SIZE` | `complete_batch_size` |
-    /// | `ZIZQ_AUTO_COMPACT_THRESHOLD` | `auto_compact_threshold` |
-    pub fn from_env() -> Result<Self, EnvConfigError> {
-        let defaults = Self::default();
-        Ok(Self {
-            cache_size: env_parse_bytes("ZIZQ_CACHE_SIZE")?.unwrap_or(defaults.cache_size),
-            data_table_size: env_parse_bytes("ZIZQ_DATA_TABLE_SIZE")?
-                .unwrap_or(defaults.data_table_size),
-            index_table_size: env_parse_bytes("ZIZQ_INDEX_TABLE_SIZE")?
-                .unwrap_or(defaults.index_table_size),
-            journal_size: env_parse_bytes("ZIZQ_JOURNAL_SIZE")?.unwrap_or(defaults.journal_size),
-            l0_threshold: env_parse("ZIZQ_L0_THRESHOLD")?.unwrap_or(defaults.l0_threshold),
-            default_completed_retention_ms: defaults.default_completed_retention_ms,
-            default_dead_retention_ms: defaults.default_dead_retention_ms,
-            default_retry_limit: defaults.default_retry_limit,
-            default_backoff: defaults.default_backoff,
-            default_commit_mode: match std::env::var("ZIZQ_DEFAULT_COMMIT_MODE").ok().as_deref() {
-                Some("fsync") => CommitMode::Fsync,
-                Some("buffered") | None => CommitMode::Buffered,
-                Some(other) => {
-                    return Err(EnvConfigError {
-                        name: "ZIZQ_DEFAULT_COMMIT_MODE".into(),
-                        value: other.into(),
-                    });
-                }
-            },
-            enqueue_commit_mode: match std::env::var("ZIZQ_ENQUEUE_COMMIT_MODE").ok().as_deref() {
-                Some("fsync") => Some(CommitMode::Fsync),
-                Some("buffered") => Some(CommitMode::Buffered),
-                None => None,
-                Some(other) => {
-                    return Err(EnvConfigError {
-                        name: "ZIZQ_ENQUEUE_COMMIT_MODE".into(),
-                        value: other.into(),
-                    });
-                }
-            },
-            enqueue_batch_size: env_parse("ZIZQ_ENQUEUE_BATCH_SIZE")?
-                .unwrap_or(defaults.enqueue_batch_size),
-            complete_batch_size: env_parse("ZIZQ_COMPLETE_BATCH_SIZE")?
-                .unwrap_or(defaults.complete_batch_size),
-            auto_compact_threshold: env_parse("ZIZQ_AUTO_COMPACT_THRESHOLD")?
-                .unwrap_or(defaults.auto_compact_threshold),
-        })
-    }
-}
-
-/// Parse an environment variable as a numeric type. Returns `Ok(None)` if
-/// unset, `Ok(Some(value))` on success, or `Err` if set but unparseable.
-fn env_parse<T: std::str::FromStr>(name: &str) -> Result<Option<T>, EnvConfigError> {
-    let val = match std::env::var(name) {
-        Ok(v) => v,
-        Err(_) => return Ok(None),
-    };
-    val.parse().map(Some).map_err(|_| EnvConfigError {
-        name: name.to_string(),
-        value: val,
-    })
-}
-
-/// Parse an environment variable as a human-readable byte size (e.g. "256MB",
-/// "1GiB", or a plain number of bytes). Returns `Ok(None)` if unset.
-fn env_parse_bytes(name: &str) -> Result<Option<u64>, EnvConfigError> {
-    let val = match std::env::var(name) {
-        Ok(v) => v,
-        Err(_) => return Ok(None),
-    };
-    val.parse::<bytesize::ByteSize>()
-        .map(|b| Some(b.as_u64()))
-        .map_err(|_| EnvConfigError {
-            name: name.to_string(),
-            value: val,
-        })
 }
 
 impl Store {
@@ -1136,6 +900,7 @@ mod tests {
     use std::collections::HashSet;
 
     use crate::store::options::{BulkDeleteOptions, EnqueueOptions, ListJobsOptions};
+    use crate::store::storage_config::StorageConfig;
     use crate::store::test_support::test_store;
     use crate::time::now_millis;
